requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
requests==2.32.3
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36

pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
joblib==1.4.2
shap==0.42.1  # <--- Compatible with NumPy 1.26

streamlit==1.40.2
python-dotenv==1.0.1
pyarrow==18.1.0
thefuzz==0.22.1  # For fuzzy matching (already used)
python-levenshtein==0.25.1  # Speeds up thefuzz
plotly==5.24.1  # For charts in Streamlit
Mako-1.3.10 
alembic-1.17.2 
colorlog-6.10.1 
optuna-4.6.0